refactor(host-application): split approval workflow services
