+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/mod.rs
@@ -1,7 +1,14 @@
+mod approval_context_service;
 mod approval_service;
 mod cleanup_plans;
+mod approval_support;
+mod builder_cleanup_service;
+mod direct_merge_workflow_service;
 mod document_service;
 mod implementation_reset_service;
+mod pull_request_provider_service;
+mod pull_request_sync_service;
+mod pull_request_workflow_service;
 mod qa_service;
 mod session_service;
 mod task_activity_guard;
